Reorganize Typescript directories
