+++ source/ts/.gitignore
@@ -1,2 +1,2 @@
-/coverage
 /node_modules
+/tests/coverage
