+++ source/ts/src/App.ts
@@ -1,4 +1,4 @@
-import * as Juce from "juce-framework-frontend";
+import Native from './Native';
 
 declare global {
   interface Window {
@@ -9,21 +9,6 @@ declare global {
     }
     App: typeof App;
   }
-}
-
-class Native {
-  canCreateMarkers = Juce.getNativeFunction("canCreateMarkers");
-  createMarkers = Juce.getNativeFunction("createMarkers");
-  getAudioSources = Juce.getNativeFunction("getAudioSources");
-  getModels = Juce.getNativeFunction("getModels");
-  getRegionSequences = Juce.getNativeFunction("getRegionSequences");
-  getTranscriptionStatus = Juce.getNativeFunction("getTranscriptionStatus");
-  getWhisperLanguages = Juce.getNativeFunction("getWhisperLanguages");
-  play = Juce.getNativeFunction("play");
-  stop = Juce.getNativeFunction("stop");
-  setPlaybackPosition = Juce.getNativeFunction("setPlaybackPosition");
-  setWebState = Juce.getNativeFunction("setWebState");
-  transcribeAudioSource = Juce.getNativeFunction("transcribeAudioSource");
 }
 
 export default class App {
