+++ source/ts/src/types/juce-framework-frontend.d.ts
