+++ source/ts/tests/__mocks__/juce-framework-frontend.js
